Working on a better data model registration.
The aim is to have the bibliography not as part of the editable document
content. The InternalList is such a note, for instance.
However, the latter is completely invisible what is not desired.
We try to resemble the same behavior (not editable) but which also gets
rendered.

Work in progress.